Add dbpf to NumImages so builder IOCs don't freeze
If acquisition is started while the PV :NumImages has value 0, the IOC
freezes and must be restarted. At some point in development history, the
PV stopped picking up its default VAL of 1, meaning that this bug would
often be encountered after starting the IOC.

This quick fix sets the VAL field to 1 immediately after IOC boot, so
the bug won't be encountered so often.